fix: Polish custom Date and Time pickers by extracting common code
