+++ app/src/main/java/org/fossasia/openevent/app/ui/views/OnDateTimeChangedListener.java
@@ -1,0 +1,7 @@
+package org.fossasia.openevent.app.ui.views;
+
+import org.fossasia.openevent.app.data.event.serializer.ObservableString;
+
+public interface OnDateTimeChangedListener {
+    void onDateChanged(ObservableString newDate);
+}
